+++ Venmo/src/test/java/com/braintreepayments/api/VenmoRequestUnitTest.java
@@ -48,6 +48,36 @@ public class VenmoRequestUnitTest {
     }
 
     @Test
+    public void getFallbackToWeb_whenTrue_returnsTrue() {
+        VenmoRequest request = new VenmoRequest(VenmoPaymentMethodUsage.SINGLE_USE);
+        request.setFallbackToWeb(true);
+        assertEquals(true, request.getFallbackToWeb());
+    }
+
+    @Test
+    public void getFallbackToWeb_whenFalse_returnsFalse() {
+        VenmoRequest request = new VenmoRequest(VenmoPaymentMethodUsage.SINGLE_USE);
+        request.setFallbackToWeb(false);
+        assertEquals(false, request.getFallbackToWeb());
+    }
+
+    @Test
+    public void getFallbackToWeb_whenNoValuePassed_defaultsToFalse() {
+        VenmoRequest request = new VenmoRequest(VenmoPaymentMethodUsage.SINGLE_USE);
+        assertEquals(false, request.getFallbackToWeb());
+    }
+
+    @Test
+    public void getIsFinalAmountAsString_returnsStringEquivalent() {
+        VenmoRequest request = new VenmoRequest(VenmoPaymentMethodUsage.MULTI_USE);
+        request.setIsFinalAmount(true);
+        assertEquals("true", request.getIsFinalAmountAsString());
+
+        request.setIsFinalAmount(false);
+        assertEquals("false", request.getIsFinalAmountAsString());
+    }
+
+    @Test
     public void parcelsCorrectly() {
         VenmoRequest request = new VenmoRequest(VenmoPaymentMethodUsage.MULTI_USE);
         request.setDisplayName("venmo-user");
@@ -60,6 +90,8 @@ public class VenmoRequestUnitTest {
         request.setDiscountAmount("2.00");
         request.setShippingAmount("1.00");
         request.setTotalAmount("10.00");
+        request.setIsFinalAmount(true);
+        request.setFallbackToWeb(true);
 
         ArrayList<VenmoLineItem> lineItems = new ArrayList<>();
         lineItems.add(new VenmoLineItem(VenmoLineItem.KIND_DEBIT, "An Item", 1, "10.00"));
@@ -83,5 +115,7 @@ public class VenmoRequestUnitTest {
         assertEquals("10.00", result.getTotalAmount());
         assertEquals(1, result.getLineItems().size());
         assertEquals("An Item", result.getLineItems().get(0).getName());
+        assertTrue(result.getIsFinalAmount());
+        assertEquals(true, result.getFallbackToWeb());
     }
 }